refactor: separate BoardsDetail logic into hook, types, and queries files
