+++ 03-maincamp-triptalk/03-CNA/src/components/boards-detail/queries.ts
@@ -1,0 +1,12 @@
+import { gql } from "@apollo/client";
+
+export const FETCH_BOARD = gql`
+  query fetchBoard($boardId: ID!) {
+    fetchBoard(boardId: $boardId) {
+      writer
+      title
+      contents
+      createdAt
+    }
+  }
+`;
