+++ 03-maincamp-triptalk/03-CNA/src/components/boards-detail/types.ts
@@ -1,0 +1,6 @@
+export interface Board {
+  writer: string;
+  title: string;
+  contents: string;
+  createdAt: string;
+}
